+++ projects/_01_ingest/ascwds/tests/fargate/test_validate_clean_ascwds_workplace.py
@@ -1,0 +1,77 @@
+import json
+import unittest
+from datetime import date
+from unittest.mock import Mock, patch
+
+import polars as pl
+
+import projects._01_ingest.ascwds.fargate.validate_clean_ascwds_workplace as job
+from utils.column_names.cleaned_data_files.ascwds_workplace_cleaned import (
+    AscwdsWorkplaceCleanedColumns as ASCWPClean,
+)
+
+PATCH_PATH = "projects._01_ingest.ascwds.fargate.validate_clean_ascwds_workplace"
+
+
+class ValidateCleanASCWDSWorkplaceTests(unittest.TestCase):
+    def setUp(self) -> None:
+        self.source_df = pl.DataFrame(
+            [
+                ("123", "20240101"),
+                ("456", "20240101"),
+                ("123", "20240201"),
+            ],
+            schema=pl.Schema(
+                [
+                    (ASCWPClean.establishment_id, pl.String),
+                    (ASCWPClean.import_date, pl.String),
+                ]
+            ),
+            orient="row",
+        )
+
+    @patch(f"{PATCH_PATH}.vl.write_reports")
+    @patch(f"{PATCH_PATH}.utils.read_parquet")
+    def test_validation_runs(
+        self,
+        mock_read_parquet: Mock,
+        mock_write_reports: Mock,
+    ):
+        mock_read_parquet.return_value = self.source_df
+        job.main("bucket", "my/source/", "my/reports/")
+
+        mock_read_parquet.assert_called_once_with(
+            source="s3://bucket/my/source/", selected_columns=job.COLS_TO_IMPORT
+        )
+        mock_write_reports.assert_called_once()
+
+    @patch(f"{PATCH_PATH}.vl.write_reports")
+    @patch(f"{PATCH_PATH}.utils.read_parquet")
+    def test_validation_report_includes_expected_validations(
+        self,
+        mock_read_parquet: Mock,
+        mock_write_reports: Mock,
+    ):
+        mock_read_parquet.return_value = self.source_df
+
+        job.main("bucket", "my/source/", "my/reports/")
+
+        validation_arg = mock_write_reports.call_args[0][0]
+        report_json = json.loads(validation_arg.get_json_report())
+
+        assertion_types_present = {item["assertion_type"] for item in report_json}
+
+        expected_assertions = {
+            "rows_distinct",
+        }
+
+        for assertion in expected_assertions:
+            self.assertIn(
+                assertion,
+                assertion_types_present,
+                f"{assertion} not found in validation report",
+            )
+
+
+if __name__ == "__main__":
+    unittest.main(warnings="ignore")