+++ projects/_01_ingest/cqc_api/fargate/utils/convert_delta_to_full.py
@@ -66,7 +66,6 @@ def convert_delta_to_full(
             merged_lf,
             full_destination,
             partition_cols=cqc_partition_keys,
-            append=False,
         )
 
         full_lf = merged_lf