+++ projects/_03_independent_cqc/_02_clean/tests/fargate/utils/clean_ascwds_filled_post_outliers/test_clean_ascwds_filled_post_outliers.py
@@ -23,6 +23,7 @@ class CleanAscwdsFilledPostOutliersTests(unittest.TestCase):
             Schemas.unfiltered_ind_cqc_schema,
             orient="row",
         )
+        self.grouped_providers_lf = pl.LazyFrame()
 
     @patch(f"{PATCH_PATH}.add_filtering_rule_column")
     @patch(
@@ -39,7 +40,14 @@ class CleanAscwdsFilledPostOutliersTests(unittest.TestCase):
         null_filled_posts_where_locations_use_invalid_missing_data_code_mock: Mock,
         add_filtering_rule_column_mock: Mock,
     ):
-        job.clean_ascwds_filled_post_outliers(self.unfiltered_ind_cqc_lf)
+        null_grouped_provders_mock.return_value = [
+            Mock(name="clean_ind_cqc_data"),
+            Mock(name="grouped_providers"),
+        ]
+
+        job.clean_ascwds_filled_post_outliers(
+            self.unfiltered_ind_cqc_lf, self.grouped_providers_lf
+        )
         winsorize_care_home_filled_posts_per_bed_ratio_outliers_mock.assert_called_once()
         null_grouped_provders_mock.assert_called_once()
         null_filled_posts_where_locations_use_invalid_missing_data_code_mock.assert_called_once()