+++ projects/_03_independent_cqc/_06_estimate_filled_posts/fargate/estimate_ind_cqc_filled_posts.py
@@ -134,7 +134,6 @@ def main(
     utils.sink_to_parquet(
         lf,
         destination,
-        append=False,
     )