+++ projects/_03_independent_cqc/_07_estimate_filled_posts_by_job_role/fargate/_04_estimate.py
@@ -61,7 +61,6 @@ def main(
         lazy_df=lf,
         output_path=estimated_data_destination,
         partition_cols=[Keys.year],
-        append=False,
     )