+++ projects/_03_independent_cqc/unittest_data/polars_ind_cqc_test_file_schemas.py
@@ -925,6 +925,7 @@ class NullGroupedProvidersSchema:
             (IndCQC.cqc_location_import_date, pl.Date()),
             (IndCQC.care_home, pl.String()),
             (IndCQC.establishment_id, pl.String()),
+            (AWPClean.nmds_id, pl.String()),
             (IndCQC.ascwds_filled_posts_dedup, pl.Float64()),
             (IndCQC.ascwds_filled_posts_dedup_clean, pl.Float64()),
             (IndCQC.number_of_beds, pl.Int64()),
@@ -999,6 +1000,26 @@ class NullGroupedProvidersSchema:
             (NGPcol.provider_pir_count, pl.Int64()),
             (NGPcol.provider_pir_sum, pl.Float64()),
             (IndCQC.ascwds_filtering_rule, pl.String()),
+        ]
+    )
+
+    select_grouped_providers_schema = pl.Schema(
+        [
+            (IndCQC.location_id, pl.String()),
+            (IndCQC.provider_id, pl.String()),
+            (IndCQC.cqc_location_import_date, pl.Date()),
+            (AWPClean.nmds_id, pl.String()),
+            (NGPcol.potential_grouped_provider, pl.Boolean()),
+            (IndCQC.ascwds_filled_posts_dedup_clean, pl.Float64()),
+        ]
+    )
+
+    final_grouped_providers_schema = pl.Schema(
+        list(select_grouped_providers_schema.items())
+        + [
+            (NGPcol.grouped_provider_status, pl.String()),
+            (NGPcol.grp_prov_identified_date, pl.Date()),
+            (NGPcol.grp_prov_fixed_date, pl.Date()),
         ]
     )