+++ projects/_07_workforce_characteristics/_01_starters_leavers_vacancies/tests/fargate/test_03_impute.py
@@ -27,5 +27,4 @@ class MainTests(unittest.TestCase):
         sink_to_parquet_mock.assert_called_once_with(
             lazy_df=ANY,
             output_path=self.IMPUTED_DATA_DESTINATION,
-            append=False,
         )