+++ utils/column_names/ind_cqc_pipeline_columns.py
@@ -398,6 +398,9 @@ class NullGroupedProviderColumns:
     potential_grouped_provider: str = "potential_grouped_provider"
     provider_pir_count: str = "provider_pir_count"
     provider_pir_sum: str = "provider_pir_sum"
+    grouped_provider_status: str = "grouped_provider_status"
+    grp_prov_identified_date: str = "grouped_provider_identified_date"
+    grp_prov_fixed_date: str = "grouped_provider_fixed_date"
 
 
 @dataclass